Move the app tour into a modal
